+++ CHANGELOG.md
@@ -19,6 +19,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - 'DEFAULT_BR_TEXT'
   - 'DEFAULT_SPAN_TEXT'
   - 'MAX_FILE_SIZE'
+- Added support for CSS combinators
+  - Added support for Child Combinator (`>`)
+  - Added support for Next Sibling Combinator (`+`)
+  - Added support for Subsequent Sibling Combinator (`~`)
 ### Changed
 - Removed automatic filtering of 'tbody' selectors (#79)
   > Remove 'tbody' from all selectors to maintain the previous state!